% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{ALELIC}
\alias{ALELIC}
\title{Allelic interactions}
\usage{
ALELIC(type = NULL, ge = NULL)
}
\arguments{
\item{type}{Type of allelic interaction. Use 'ad' for additivity, 'dom'
for complete dominance, 'domp' for partial dominance and 'sob' for
overdominance.}

\item{ge}{Type of GxE interaction. Use 'aus' for no interaction,
'simple' for simple interaction and 'complex' for complex interaction.}
}
\description{
Examples of allelic and gene interactions
}
\author{
Willyan Junior Adorian Bandeira

Ivan Ricardo Carvalo

Murilo Vieira Loro

Leonardo Cesar Pradebon

Jose Antonio Gonzalez da Silva
}
