% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{ALELICA}
\alias{ALELICA}
\title{Interações alélicas}
\usage{
ALELICA(interactions, type = NULL, ge = NULL)
}
\description{
Exemplos didáticos de interações alélicas e gênicas
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
