% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiental.R
\name{atsum}
\alias{atsum}
\title{Accumulated Thermal Sum}
\usage{
atsum(TMED, crop = "maize", plot = F)
}
\arguments{
\item{TMED}{The column with the average air temperature values}

\item{crop}{Parameter to define the culture. Use "maize", "soybean", "flax",
"trit" or "oat"}

\item{plot}{Logical argument. Plot a graph of thermal accumulation if TRUE.}
}
\description{
Calculates the daily and accumulated thermal sum of crops
}
\examples{
\donttest{
library(EstimateBreed)

data("clima")
clima <- clima(1:150,)

with(clima,atsum(TMED,crop="maize"))
}
}
\author{
Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalo

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva
}
