% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{aveia}
\alias{aveia}
\title{Dataset: Oat data}
\format{
A data.frame with 54 observations and 6 variables:
\describe{
  \item{GEN}{14 white oat genotypes.}
  \item{BLOCO}{Experiment blocks.}
  \item{NG2M}{Number of grains larger than 2 mm.}
  \item{MG}{Grain mass}
  \item{MC}{Caryopsis dough}
  \item{RG}{Grain yield (in kg per ha)}
}
}
\source{
Real field data for use.
}
\usage{
aveia
}
\description{
Data set with oat genotypes and industry variables.
}
\keyword{datasets}
