% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiental.R
\name{balanco_hidrico_complexo_auto}
\alias{balanco_hidrico_complexo_auto}
\title{Balanço hídrico}
\usage{
balanco_hidrico_complexo_auto(
  P,
  ET,
  R,
  D,
  I,
  lat,
  lon,
  altitude,
  S_max,
  porosidade,
  capacidade_infiltracao,
  data_inicio,
  data_fim
)
}
\description{
Balanço hídrico
}
