% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{blupis}
\alias{blupis}
\title{Número de indivíduos a serem selecionados em cada família}
\usage{
blupis()
}
\arguments{
\item{y}{Variável resposta}

\item{r}{A coluna com o efeito de repetição (fixos)}

\item{p}{A couna com o efeito de parcela (aleatório)}

\item{f}{A coluna com o efeito de genético de dominância associado a famílias
de irmãos-completos (aleatório)}

\item{b}{A coluna com o efeito dos blocos incompletos (aleatório)}

\item{A}{Matriz de parentesco obtida com o AGHmatrix}
}
\description{
Definição do número de indivíduos a serem selecionados pelo BLUPis
}
\references{
Oliveira, R. A., Daros, E., Resende, M. D. V., Bespalhok-Filho, J. C.,
Zambon, J. L. C., Souza, T. R., & Lucius, A. S. F. (2011). Procedimento
Blupis e seleção massal em cana-de-açúcar. Bragantia, 70(4), 796–800.
Model 35 (SELEGEN)
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
