% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clima.R
\docType{data}
\name{clima}
\alias{clima}
\title{Conjunto de Dados Climáticos para Predições}
\format{
Um data.frame com 8760 observações e 5 variáveis:
\describe{
  \item{MO}{Mês do ano.}
  \item{DY}{Dia do ano.}
  \item{HR}{Hora do dia.}
  \item{TMED}{Temperatura Média do Ar - em ºC.}
  \item{RH}{Umidade Relativa do ar - em \%.}
}
}
\source{
Dados obtidos na plataforma Nasa Power (https://power.larc.nasa.gov/).
}
\usage{
clima
}
\description{
Dados de temperatura média do ar e umidade relativa do ar para o período de
um ano, com hora, dia e mês.
}
\keyword{datasets}
