% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefend.R
\docType{data}
\name{coefend}
\alias{coefend}
\title{Conjunto de Dados para Coef. de Endogamia}
\format{
Um data.frame com 7 observações e 5 variáveis:
\describe{
  \item{Var}{Variável.}
  \item{VF}{Variância fenotípica.}
  \item{VG}{Variância genética.}
  \item{h}{Herdabilidade}
  \item{DS}{Diferencial de seleção.}
}
}
\source{
Dados reais de campo para utilização.
}
\usage{
coefend
}
\description{
Conjunto de dados de variância fenotípica, genotípica, herdabilidade e
diferencial de seleção para diferentes variáveis.
}
\keyword{datasets}
