% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{COI}
\alias{COI}
\title{Inbreeding coefficient}
\usage{
COI(var, VG, VF, generation = "all")
}
\arguments{
\item{var}{Column with the variable name}

\item{VG}{Column with genotypic variance}

\item{VF}{Column with phenotypic variance}
}
\value{
Returns the total, additive and dominance variance values based on
the variance components for a given variable.
}
\description{
Function for calculating the inbreeding coefficient
}
\examples{
\donttest{
library(EstimateBreed)

var <- c("A","B","C","D","E")
VG <- c(2.5, 3.0, 2.8, 3.2, 2.7)
VF <- c(1.2, 1.5, 1.3, 1.6, 1.4)
data <- data.frame(var,VG,VF)

#Calculating for all generations
with(data,COI(var,VG,VF,geracao = "all"))

Calculating for just one generation
with(data,COI(var,VG,VF,geracao = "F3"))
}
}
\author{
Willyan Junior Adorian Bandeira

Ivan Ricardo Carvalo

Murilo Vieira Loro

Leonardo Cesar Pradebon

Jose Antonio Gonzalez da Silva
}
