% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restr.R
\name{cvar}
\alias{cvar}
\title{Estimativa dos componentes de variância pela restrição da variabilidade das
testemunhas.}
\usage{
cvar(GEN, REP, Xi, approach = NULL)
}
\arguments{
\item{GEN}{A coluna com o nome dos genótipos (sem testemunhas).}

\item{REP}{A coluna com as repetições (se houver).}

\item{Xi}{A coluna com o valor observado para determinado genótipo.}

\item{approach}{Método a ser utilizado para estimativa dos componentes de
variância. Usar "apI" para regressão genitor progênie, "apII" para o método
da soma de quadrados de blocos aumentados com testemunhas intercalares, "apIII"
para o método com modelos lineares mistos com efeitos genéticos aleatórios.}
}
\value{
Retorna uma tabela com os genótipos e os índices selecionados.
Quanto maior o valor do índice, mais resiliente é o genótipo.
}
\description{
Método para restrição da variabilidade das testemunhas proposta por
Carvalho et al. (2023). Utiliza a restriçõa da média mais ou menos um desvio
padrão, o que restringe a variação a partir da remoção de valores assimétricos.
}
\references{
Carvalho, I. R., Silva, J. A. G. da, Moura, N. B., Ferreira, L. L.,
Lautenchleger, F., & Souza, V. Q. de. (2023). Methods for estimation of
genetic parameters in soybeans: An alternative to adjust residual variability.
Acta Scientiarum. Agronomy, 45, e56156.
https://doi.org/10.4025/actasciagron.v45i1.56156
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
