% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{default_seg}
\alias{default_seg}
\title{Standard Segregation}
\usage{
default_seg(MELHORAMENTO)
}
\arguments{
\item{MELHORAMENTO}{Base parameter for the print table function}
}
\description{
Didactic table of standard segregation by generation
}
\author{
Willyan Junior Adorian Bandeira

Ivan Ricardo Carvalo

Murilo Vieira Loro

Leonardo Cesar Pradebon

Jose Antonio Gonzalez da Silva
}
