% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiental.R
\name{deltat}
\alias{deltat}
\title{Aplicação de defensivos agrícolas}
\usage{
deltat(LON, LAT, type = 1, days = 7, details = FALSE)
}
\arguments{
\item{LON}{Longitude}

\item{LAT}{Latitude}

\item{type}{Tipo de análise. Utilize 1 para forecast e 2 para dados temporais.}

\item{days}{Número de dias}
}
\description{
Determinação do momento ideal para aplicação
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
