% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isgr.R
\name{desv_clim}
\alias{desv_clim}
\title{Auxiliary function for calculating ISGR}
\usage{
desv_clim(ENV, TMED, PREC)
}
\arguments{
\item{ENV}{Identification of each selection environment (to differentiate if
there is more than one cultivation cycle).}

\item{TMED}{Average air temperature (in ºC) during the cycle in each environment.}

\item{PREC}{Rainfall (in mm) during the cultivation cycle in each environment}
}
\value{
A dataframe containing the identifier of the selection environment and
the standard deviations for temperature and precipitation.
}
\description{
This function receives a dataframe with temperature and precipitation data
and calculates the standard deviation of these parameters for each environment.
}
\examples{
\donttest{
library(EstimateBreed)
data("desvamb")
head(desvamb)

with(desvamb,desv_clim(ENV,TMED,PREC))
}
}
\author{
Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva
}
