% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desvamb.R
\docType{data}
\name{desvamb}
\alias{desvamb}
\title{Conjunto de Dados para Desvio Ambiental}
\format{
Um data.frame com 449 observações e 3 variáveis:
\describe{
  \item{ENV}{Ambiente de seleção.}
  \item{TMED}{Temperatura média do ar (em ºC).}
  \item{PREC}{Precpitação pluviométrica.}
}
}
\source{
Dados reais de campo para utilização.
}
\usage{
desvamb
}
\description{
Conjunto de dados com genótipos de aveia e variáveis de indústria.
}
\keyword{datasets}
