% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defh.R
\name{dhid}
\alias{dhid}
\title{Determinação do Déficit Hídrico}
\usage{
dhid(CICLO, FASE, ETo, AD = 4, PROF = 100, Cultura = "Soja")
}
\arguments{
\item{CICLO}{Coluna referente ao ciclo da cultura}

\item{FASE}{Fase em que a cultura se encontra}

\item{ETo}{Evapotranspiração potencial do dia}

\item{PROF}{Profundidade Aproximada do Solo (em cm)}

\item{Cultura}{Definir a cultura (Padrão = "Soja")}

\item{ADP}{Água Disponível Predita}

\item{plot}{Plotar um gráfico do acúmulo (Padrão é F (FALSE))}
}
\description{
Estimativa do déficit hídrico das culturas agrícolas
}
