% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{didint}
\alias{didint}
\title{Allelic and genotype-environment interactions}
\usage{
didint(type = NULL, ge = NULL)
}
\arguments{
\item{type}{Type of allelic interaction. Use 'ad' for additivity, 'dom'
for complete dominance, 'domp' for partial dominance and 'sob' for
overdominance.}

\item{ge}{Type of GxE interaction. Use 'aus' for no interaction,
'simple' for simple interaction and 'complex' for complex interaction.}
}
\value{
Plot graphs representing allelic and genotype x environment
interactions.
}
\description{
Didactic function - Examples of allelic and gene interactions
}
\examples{
library(EstimateBreed)

didint (type="ad")
didint (type="dom")
didint (type="domp")
didint (type="sob")

didint (ge="aus")
didint (ge="simple")
didint (ge="complex")
}
\author{
Willyan Junior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

Jose Antonio Gonzalez da Silva
}
