% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estresse.R
\name{estresse}
\alias{estresse}
\title{índices de Estresse para Seleção}
\usage{
estresse(
  GEN,
  YS,
  YC,
  index = "ALL",
  bygen = T,
  plot = F,
  xlab = "Genótipo",
  ylab = "Valores",
  ...
)
}
\arguments{
\item{GEN}{Genótipos a serem selecionados.}

\item{YS}{Produtividade do genótipo sem condições estressoras.}

\item{YC}{Prdotuvidade do genótipo sob condições estressoras.}

\item{index}{Índice a ser calculado (Padrão "ALL"). Os índices a serem utilizados
são: "STI" - Stress Tolerance Index, "YI" - Yield Index, "GMP" - Geometric Mean Productivity,
"MP" - Mean Productivity, "MH" - Harmonic Mean, "SSI" - Stress Stability Index,
"YSI" - Yield Stability Index, "RSI" - Relative Stress Index.}

\item{bygen}{Retorna a média de cada genótipo se "TRUE". Somente desta forma
será possível plotar gráficos.}

\item{plot}{Plotar gráfico se igual a "TRUE" (Padrão "F").}

\item{xlab}{Ajustar o título do eixo x no gráfico.}

\item{ylab}{Ajustar o título do eixo y no gráfico.}
}
\value{
Retorna uma tabela com os genótipos e os índices selecionados.
Quanto maior o valor do índice, mais resiliente é o genótipo.
}
\description{
Índices de seleção de genótipos conduzidos sob condições de estresse citados
por Ghazvini et al. (2024).
}
\examples{
\donttest{
library(Breeding)

data("aveia")
with(aveia,estresse(GEN,MC,MG,index = "ALL",bygen=T))

}
}
\references{
Ghazvini, H., Pour-Aboughadareh, A., Jasemi, S.S. et al.
A Framework for Selection of High-Yielding and Drought-tolerant
Genotypes of Barley: Applying Yield-Based Indices and Multi-index
Selection Models. Journal of Crop Health 76, 601–616 (2024).
https://doi.org/10.1007/s10343-024-00981-1
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho, Murilo V. Loro,
Leonardo C. Pradebon, José A. G. da Silva
}
