% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiental.R
\name{fototermal}
\alias{fototermal}
\title{Determinação do índice fototermal}
\usage{
fototermal(DIA, TMED, N, plot = F)
}
\arguments{
\item{DIA}{A coluna com o dia de ciclo}

\item{TMED}{A coluna com os valores de temperatura média}

\item{N}{A coluna com os valores de fotoperíodo}

\item{plot}{Argumento lógico. Plotar um gráfico se TRUE}
}
\description{
Cálculo do índice fototermal
}
