% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genot.R
\docType{data}
\name{genot}
\alias{genot}
\title{Conjunto de Dados GXE}
\format{
Um data.frame com 55 observações e 5 variáveis:
\describe{
  \item{GEN}{Linhagens selecionadas em um experimento GXE.}
  \item{ENV}{Ambientes de seleção.}
  \item{NG}{Número de grãos mensurado nas linhagens.}
  \item{MG}{Massa de grãos mensurada nas linhagens (e
  m g)}
  \item{CICLO}{Duração do ciclo de cultivo (em dias)}
}
}
\source{
Dados reais de campo para utilização.
}
\usage{
genot
}
\description{
Conjunto de dados com linhagens e testemnunhas de um experimento GxE.
}
\keyword{datasets}
