% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{genot2}
\alias{genot2}
\title{Data Set for obtaining genetic parameters.}
\format{
A data.frame with 60 observations and 4 variables:
\describe{
  \item{Pop}{Column with population names.}
  \item{Gen}{Column with genotype names.}
  \item{Rep}{Column with replications.}
  \item{VAR1}{Column with numerical values of the random variable.}
}
}
\source{
Simulated data for use.
}
\usage{
genot2
}
\description{
Dataset with two breeding populations, 20 genotypes per population and three
 replicates per genotype.
}
\keyword{datasets}
