% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{genpar}
\alias{genpar}
\title{Parâmetros gerais para seleção}
\usage{
genpar(POP, GEN, REP = NULL, vars, K = 0.05, type = "balanced", check = FALSE)
}
\arguments{
\item{POP}{A coluna com a população em melhoramento.}

\item{GEN}{A coluna com os genótipos selecionados dentro da população.}

\item{REP}{A coluna com as repetições (se houver).}

\item{K}{Pressão de seleção (Padrão 0.05).}

\item{type}{Tipo de experimento (balanceado ou desbalanceado). Utilize
"balanced" para balanceado e "unb" para desbalanceado.}

\item{check}{Argumento lógico. Faz as verificações do pressuposto do modelo
estatístico se o valor for igual a TRUE.}

\item{VAR}{A coluna com a variável de interesse.}
}
\description{
Função para determinar parâmetros de seleção, com base em um experimento
realizado na cultura do arroz. Destinada para avaliação isolada do desempenho
de linhagens dentro de uma determinada população.
}
\references{
Yadav, S. P. S., Bhandari, S., Ghimire, N. P., Mehata, D. K., Majhi, S. K.,
Bhattarai, S., Shrestha, S., Yadav, B., Chaudhary, P., & Bhujel, S. (2024).
Genetic variability, character association, path coefficient, and diversity
analysis of rice (Oryza sativa L.) genotypes based on agro-morphological
traits. International Journal of Agronomy, 2024, Article ID 9946332.
https://doi.org/10.1155/2024/9946332
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
