% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{GS}
\alias{GS}
\title{Selection pressure}
\usage{
GS(Var, h, VF, P = "1")
}
\arguments{
\item{Var}{The column with the name of the variables of interest}

\item{h}{The column with the restricted heritability values}

\item{VF}{The column with the phenotypic variance values}

\item{P}{The column with the values observed for the progenies}
}
\description{
Response to selection weighted by selection pressure
}
\author{
Willyan Junior Adorian Bandeira

Ivan Ricardo Carvalo

Murilo Vieira Loro

Leonardo Cesar Pradebon

Jose Antonio Gonzalez da Silva
}
