% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{GS2}
\alias{GS2}
\title{Diferencial de Seleção Simples}
\usage{
GS2(Var, h, DS)
}
\arguments{
\item{Var}{A coluna com o nome das variáveis de interesse}

\item{h}{A coluna com os valores de herdabilidade no sentido restrito}

\item{DS}{A coluna com o valor do diferencial de seleção a ser aplicado para
cada variável}
}
\description{
Resposta a seleção ponderada pelo Diferencial de Seleção Simples
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho, Murilo V. Loro,
Leonardo C. Pradebon, José A. G. da Silva
}
