% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{GS3}
\alias{GS3}
\title{Response to Selection by Control of Genitors}
\usage{
GS3(Var, h, VF, P = "1")
}
\arguments{
\item{Var}{The column with the variables of interest}

\item{h}{The column with the restricted heritability values}

\item{VF}{The column with the phenotypic variance values}

\item{P}{The column with the progeny values}
}
\description{
Consider knowing only the maternal parent, without controlling the pollinator,
or direct selection without parents
}
\author{
Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva
}
