% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{GS4}
\alias{GS4}
\title{Response to Selection by Year}
\usage{
GS4(Var, h, VF, P = "1", Ano)
}
\arguments{
\item{Var}{The column with the variables of interest}

\item{h}{The column with the restricted heritability values}

\item{VF}{The column with the phenotypic variance values}

\item{P}{The column with the value obtained for the progenies}

\item{Year}{The column with the year of selection}
}
\description{
Response to selection weighted by Pressure Weighted by Year of selection and
Year
}
\author{
Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva
}
