% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indgerm.R
\name{gvri}
\alias{gvri}
\title{Selection for Grain Volume}
\usage{
gvri(
  GEN,
  C,
  L,
  E,
  stat = "everything",
  plot = F,
  ylab = "GVRI",
  xlab = "Genotype"
)
}
\arguments{
\item{GEN}{The column with the genotype name}

\item{C}{Grain length}

\item{L}{Grain width}

\item{E}{Grain thickness}

\item{stat}{Extract or not the average per genotype. Use 'everything' to obtain
information on all the observations or 'mean' to extract the average.}

\item{plot}{Logical argument. Plot a graph if TRUE}

\item{ylab}{Y axis name}

\item{xlab}{X axis name}
}
\description{
Calculation of the selection index for grain volume, based on the values for
grain length, width and thickness
}
\examples{
\donttest{
library(EstimateBreed)

Gen = rep(paste0("G", 1:10), each = 3)
Rep = rep(1:3, times = 10)
L = round(rnorm(30, mean = 3.2, sd = 0.3), 2)
C = round(rnorm(30, mean = 8.5, sd = 0.5), 2)
E = round(rnorm(30, mean = 2.1, sd = 0.2), 2)

data <- data.frame(Gen,Rep,L,C,E)

with(data,gvri(Gen,C,L,E, stat="mean", plot=TRUE))
}
}
\references{
Carvalho, I. R., de Pelegrin, A. J., Szareski, V. J., Ferrari, M., da Rosa, T.
C., Martins, T. S., dos Santos, N. L., Nardino, M., de Souza, V. Q., de
Oliveira, A. C., & da Maia, L. C. (2017). Diallel and prediction (REML/BLUP)
for yield components in intervarietal maize hybrids. Genetics and Molecular
Research, 16(3), gmr16039734.
 https://doi.org/10.4238/gmr16039734
}
\author{
Willyan Junior Adorian Bandeira

Ivan Ricardo Carvalo

Murilo Vieira Loro

Leonardo Cesar Pradebon

Jose Antonio Gonzalez da Silva
}
