% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterose.R
\name{heterose}
\alias{heterose}
\title{Função para cálculo da heterose e heterobeltiose}
\usage{
heterose(GEN, GM, GP, PR, REP, param = "all")
}
\arguments{
\item{GEN}{A coluna com o nome do genótipo}

\item{GM}{A coluna com a média do genitor materno}

\item{GP}{A coluna com a média do genitor paterno}

\item{PR}{A coluna com a média da progênie}

\item{REP}{A coluna com as repetições}

\item{param}{Valor para determinar o parâmetro a ser calculado. Padrão é "all".
Para calcular apenas a heterose, utilize "het". Para calcular apenas a
heterobeltiose, utilize "hetb".}
}
\description{
Cálculo dos parâmetros da heterose e heterobeltiose de híbridos de milho
}
\examples{
\donttest{
library(Breeding)

data("maize")
#Extrair heterose e heterobeltiose
with(maize,heterose(HIB,GM,GP,P,stat="all"))

#Extrair apenas heterose
with(maize,heterose(HIB,GM,GP,P,param = "het"))

#Extrair apenas heterobeltiose
with(maize,heterose(HIB,GM,GP,P,param = "hetb"))
}
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho, Murilo V. Loro,
Leonardo C. Pradebon, José A. G. da Silva
}
