% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indgerm.R
\name{index_vigor}
\alias{index_vigor}
\title{Índice de vigor multivariado em sementes - COMPLETO}
\usage{
index_vigor(GEN, PC, G, CPA, RAD, MS, EC)
}
\arguments{
\item{PC}{desc}

\item{G}{desc}

\item{CPA}{desc}

\item{RAD}{desc}

\item{MS}{desc}

\item{EC}{desc}
}
\description{
Determinação do vigor multivariado de sementes
}
\references{
Szareski, V. J., Carvalho, I. R., Demari, G. H., Rosa, T. C. D.,
Souza, V. Q. D., Villela, F. A.,Aumonde, T. Z. (2018).
Multivariate index of soybean seed vigor:
a new biometric approach applied to the effects of genotypes and
environments. Journal of Seed Science, 40(4), 396-406.
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho, Murilo V. Loro,
Leonardo C. Pradebon, José A. G. da Silva
}
