% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indgerarroz.R
\name{indger}
\alias{indger}
\title{Índice de germinação pela contagem subsequente}
\usage{
indger(TESTE, DIA, TSG, NST)
}
\arguments{
\item{TESTE}{Número identificador do teste realizdo}

\item{DIA}{Valores numéricos para os dias testados}

\item{TSG}{Nome da coluna com o total de sementes germinadas em cada período}

\item{NST}{Nome da coluna padrão com o número de sementes testadas}
}
\description{
Estimativa do índice de germinação pela contagem subsequente de plântulas
germinadas em determinado período (Wang et al., 2017)
}
\examples{
with(data,indviab(genot,var1,var2))
}
\references{
Wang, Pan, Li, Dong, Wang, Li-jun and Adhikari, Benu. "Effect of High
Temperature Intermittent Drying on Rice Seed Viability and Vigor" International
Journal of Food Engineering, vol. 13, no. 10, 2017, pp. 20160433.
https://doi.org/10.1515/ijfe-2016-043
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
