% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indqi.R
\name{indviab}
\alias{indviab}
\title{Estimativa de índices de espiga}
\usage{
indviab(
  GEN,
  var1,
  var2,
  ylab = "Índice",
  xlab = "Genótipo",
  stat = "all",
  plot = F
)
}
\arguments{
\item{var1}{Nome da coluna que contém a primeira variável}

\item{var2}{Nome da coluna que contém a segunda variável}

\item{genot}{Nome da coluna que contém os genótipos}
}
\description{
Estimativa do índice de viabilidade a partir de duas variáveis de campo.
}
\examples{
library(Breeding)

data("trigo")
#Índice de viabilidade de espiga
with(trigo,indviab(TEST,NGE,NNE))

#Índice de colheita da espiga
with(trigo,indviab(TEST,MGE,ME))

#Índice de deposição de espiguetas na espiga
#'with(trigo,indviab(TEST,NEE,CE))
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
