% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indqi.R
\name{is_ptnerg}
\alias{is_ptnerg}
\title{Selection index for protein and grain yield}
\usage{
is_ptnerg(GEN, PTN, RG)
}
\arguments{
\item{GEN}{The column with the name of the genotype}

\item{PTN}{The column with the crude protein values}

\item{RG}{The column with the grain yield values (in kg per ha)}
}
\description{
Selection index for protein and grain yield (Pelegrin et al., 2017).
}
\references{
de Pelegrin, A. J., Carvalho, I. R., Nunes, A. C. P., Demari, G. H., Szareski,
V. J., Barbosa, M. H., ... & da Maia, L. C. (2017).
Adaptability, stability and multivariate selection by mixed models.
American Journal of Plant Sciences, 8(13), 3324.
}
\author{
Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva
}
