% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indqi.R
\name{is_qindustrial}
\alias{is_qindustrial}
\title{Estimativa do índice de seleção para qualidade industrial do trigo}
\usage{
is_qindustrial(GEN, NQ, W, PTN)
}
\arguments{
\item{GEN}{A coluna com o nome do genótipo}

\item{NQ}{A coluna com o valor do falling number}

\item{W}{A coluna com a força de glúten (W)}

\item{PTN}{A coluna com os valores de proteína}
}
\description{
Função para determinar índices de qualidade industrial de genótipos de trigo,
descritas por Szareski et al. (2019).
}
\examples{
\donttest{
library(Breeding)

data("ptn")
with(ptn,is_qindustrial(Cult,NQ,W,PTN))
}
}
\references{
Szareski, V. J., Carvalho, I. R., Kehl, K., Levien, A. M.,
Lautenchleger, F., Barbosa, M. H., ... & Aumonde, T. Z. (2019).
Genetic and phenotypic multi-character approach applied to multivariate
models for wheat industrial quality analysis.
Genetics and Molecular Research, 18(3), 1-14.
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
