% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isgr.R
\name{isgr}
\alias{isgr}
\title{ISGR - Índice de Seleção Genético para Resiliência.}
\usage{
isgr(GEN, ENV, NG, MG, CICLO, req = 3.5)
}
\arguments{
\item{GEN}{Coluna referente aos genótipos. As linhagens devem obrigatoriamente
ter o prefixo "L" antes do número. Ex: L139.}

\item{ENV}{Coluna referente ao ambiente de seleção.}

\item{NG}{Número de grãos de todos os genótipos avaliados}

\item{MG}{Massa de grãos de todos os genótipos avaliados}

\item{CICLO}{Número de dias do ciclo, para definir a precipitação pluviométrica
ideal (valor de 3.5 mm por dia)}

\item{req}{Valor da demanda hídrica diária média para a cultura da soja (padrão 3.5 mm)}
}
\description{
Estimativa do índice de seleção para resiliência ambiental (Bandeira et al., 2024).
}
\examples{
\donttest{
library(Breeding)
#Obter os desvios ambientais
data("desvamb")
with(desvamb, desv_clim(ENV, TMED, PREC))

#Calcular o ISGR
data("genot")
with(genot, isgr(GEN, ENV, NG, MG, CICLO))
}
}
\references{
Bandeira, W. J. A., Carvalho, I. R., Loro, M. V., da Silva, J. A. G.,
Dalla Roza, J. P., Scarton, V. D. B., Bruinsma, G. M. W., & Pradebon, L. C. (2024).
Identifying soybean progenies with high grain productivity and stress resilience
to abiotic stresses. Aust J Crop Sci, 18(12), 825-830.
https://doi.org/10.21475/ajcs.24.18.12.p98
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
