% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estresse.R
\name{itu}
\alias{itu}
\title{Índice de Temperatura e Umidade}
\usage{
itu(CICLO, TM, UR)
}
\arguments{
\item{CICLO}{Coluna referente ao Ciclo}

\item{TM}{Colune referente à Temperatura Média}

\item{UR}{Coluna referente à Umidade Relativa}
}
\description{
Cálculo do Índice de Temperatura e Umidade (ITU)
}
