% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indgerm.R
\name{ivig_simp}
\alias{ivig_simp}
\title{Índice de vigor multivariado em sementes - SIMPLES}
\usage{
ivig_simp(GEN, PC, G)
}
\arguments{
\item{PC}{desc}

\item{G}{desc}
}
\description{
Índice de vigor simples de sementes (Szareski et al., 2018).
}
\references{
Szareski, V. J., Carvalho, I. R., Kehl, K., Levien, A. M., Nardino,
M., Dellagostin, S. M., ... & Aumonde, T. Z. (2018).
Adaptability and stability of wheat genotypes according to the
phenotypic index of seed vigor. Pesquisa Agropecuária Brasileira,
53, 727-735.
}
