% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaf.R
\name{lai}
\alias{lai}
\title{Leaf Area Index (LAI)}
\usage{
lai(GEN, W, L, TNL, TDL, crop = "soy", sp = 0.45, sden = 14)
}
\arguments{
\item{GEN}{The column with the genotype name}

\item{W}{The column with the width of the leaf.}

\item{L}{The column with the length of the leaf.}

\item{TNL}{The column with the total number of leaves.}

\item{TDL}{The column with the total number of dry leaves.}

\item{crop}{Crop sampled. Use “soy” for soybean and “maize” for corn, “trit”
for wheat, “rice” for rice, “bean” for bean, “sunflower” for sunflower,
“cotton” for cotton, “sugarcane” for sugarcane, “potato” for potato and
“tomato” for tomato.}

\item{sp}{Row spacing (Standard sp=0.45).}

\item{sden}{Sowing density, in plants per linear meter (standard sden=14).}
}
\description{
Utility function for estimating crop LAI
}
\references{
Meira, D., Queiróz de Souza, V., Carvalho, I. R., Nardino, M., Follmann,
D. N., Meier, C., Brezolin, P., Ferrari, M., & Pelegrin, A. J. (2015).
Plastocrono e caracteres morfológicos da soja com hábito de crescimento
indeterminado. Revista Cultivando o Saber, 8(2), 184-200.
}
\author{
Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva
}
