% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{leafarea}
\alias{leafarea}
\title{Data Set for Leaf Area Index}
\format{
A data.frame with 10 observations and 5 variables:
\describe{
  \item{GEN}{Column with the genotypes.}
  \item{C}{Leaf lenght}
  \item{L}{Leaf width}
  \item{TNL}{Total number of leaves.}
  \item{TDL}{Total dry leavesh.}
}
}
\source{
Simulated data.
}
\usage{
leafarea
}
\description{
Data set with 10 genotypes and values for leaf length, leaf width, number
of total leaves and number of dry leaves
}
\keyword{datasets}
