% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin.R
\docType{data}
\name{lin}
\alias{lin}
\title{Conjunto de Dados de Trigo com Proteína e RG}
\format{
Um data.frame com 24 observações e 7 variáveis:
\describe{
  \item{POP}{População base.}
  \item{MGP_MF}{Média fenotípica da massa de grãos por planta.}
  \item{MGP_GP}{Média genotípica da messa de grãos por planta.}
  \item{VF}{Variância fenotípica}
  \item{VG}{Variância genética}
  \item{H2}{Herdabilidade no sentido amplo}
  \item{Test}{Parâmetros das testemunhas}
}
}
\source{
Dados reais de campo para utilização.
}
\usage{
lin
}
\description{
Conjunto de dados com genótipos de trigo, percentual de proteína e rendimento
de grãos.
}
\keyword{datasets}
