% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearest.R
\name{linearest}
\alias{linearest}
\title{Estimates using polynomial equations.}
\usage{
linearest(indep, dep, type = NULL, alpha = 0.05)
}
\arguments{
\item{indep}{Name of the column with the independent variable.}

\item{dep}{Name of the dependent variable column}

\item{type}{Type of analysis to be carried out. Use 'MET' to extract the
maximum technical efficiency or 'plateau' for plateau regression.}

\item{alpha}{Significance of the test.}
}
\description{
Determination of maximum technical efficiency (MET) and plateau regression.
}
\examples{
\donttest{
library(EstimateBreed)
data("mtcars")

with(mtcars,linearest(wt,mpg,type = "MET"))
}
}
\author{
Willyan Junior Adorian Bandeira

Ivan Ricardo Carvalo

Murilo Vieira Loro

Leonardo Cesar Pradebon

Jose Antonio Gonzalez da Silva
}
