% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maize.R
\docType{data}
\name{maize}
\alias{maize}
\title{Conjunto de Dados de Milho}
\format{
Um data.frame com 4 observações e 3 variáveis:
\describe{
  \item{P}{Progênies.}
  \item{GM}{Genitor Materno.}
  \item{GP}{Genitor Paterno.}
}
}
\source{
Dados simulados.
}
\usage{
maize
}
\description{
Conjunto de dados com progênies e genitores maternos e paternos de milho.
}
\keyword{datasets}
