% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutagenico.R
\name{mut_index}
\alias{mut_index}
\title{Índice de vigor multivariado em sementes}
\usage{
mut_index(mut, MSG, MST, GT, DT, SL)
}
\arguments{
\item{mut}{Método de mutação}

\item{MSG}{Média de Sementes Germinadas}

\item{MST}{Média de Sementes Total}

\item{GT}{Número de plântulas por dia durante o tempo t}

\item{DT}{Número de Dias Avaliados}

\item{SL}{Comprimento da plântula}
}
\description{
Citação: "A new method for mutation inducing in rice
by using DC electrophoresis bath and its mutagenic ef
}
\references{
Zou, M., Tong, S., Zou, T. et al. A new method for mutation inducing in rice
by using DC electrophoresis bath and its mutagenic effects. Sci Rep 13, 6707
(2023). https://doi.org/10.1038/s41598-023-33742-7
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho, Murilo V. Loro,
Leonardo C. Pradebon, José A. G. da Silva
}
