% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.R
\docType{data}
\name{pheno}
\alias{pheno}
\title{Conjunto de Dados para Estimativa do Plastocrono da Soja}
\format{
Um data.frame com 135 observações e 5 variáveis:
\describe{
  \item{CICLO}{Dias do ciclo da soja.}
  \item{GEN}{A coluna com o nome do genótipo.}
  \item{TMED}{A coluna com os valores de temperatura média.}
  \item{EST}{A coluna com o estádio fenológico.}
  \item{NN}{A coluna com o número de nós.}
}
}
\source{
Dados fictícios para utilização.
}
\usage{
pheno
}
\description{
Conjunto de dados fictícios para estimativa do plastocrono da soja com base
no número de nós
}
\keyword{datasets}
