% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiental.R
\name{plastocrono}
\alias{plastocrono}
\title{Estimativa do plastocrono da soja}
\usage{
plastocrono(GEN, TMED, NNOS, habito = "ind", plot = F)
}
\arguments{
\item{GEN}{Coluna referente ao genótipo}

\item{TMED}{Coluna referente a temperatura média}

\item{NNOS}{Número de nós mensurados a campo}

\item{habito}{Hábito de crescimento do genótipo (padrão="ind")}

\item{plot}{Imprimir o gráfico (padrão=T)}
}
\description{
Estimativa do plastocrono da soja por meio da Tmed
}
