% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptn.R
\docType{data}
\name{ptn}
\alias{ptn}
\title{Conjunto de Dados de Milho}
\format{
Um data.frame com 360 observações e 5 variáveis:
\describe{
  \item{Cult}{Cultivares de trigo.}
  \item{Am}{Número identificador da amostra.}
  \item{NQ}{Falling Number.}
  \item{W}{Força de Glúten (W).}
  \item{PTN}{Proteína.}
}
}
\source{
Dados reais de laboratório.
}
\usage{
ptn
}
\description{
Conjunto de dados com progênies e genitores maternos e paternos de milho.
}
\keyword{datasets}
