% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptnrg.R
\docType{data}
\name{ptnrg}
\alias{ptnrg}
\title{Conjunto de Dados de Trigo}
\format{
Um data.frame com 360 observações e 5 variáveis:
\describe{
  \item{CULTIVAR}{Cultivares de trigo.}
  \item{REP}{Número da repetição.}
  \item{PTN}{Proteína.}
  \item{RG}{Rendimento de grãos (kg ha)}
}
}
\source{
Dados reais de campo.
}
\usage{
ptnrg
}
\description{
Conjunto de dados com progênies e genitores maternos e paternos de milho.
}
\keyword{datasets}
