% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{reg_GP}
\alias{reg_GP}
\title{Regression Genitor Progeny}
\usage{
reg_GP(ind, Genitor, Progenie)
}
\arguments{
\item{ind}{description}

\item{Genitor}{description}

\item{Progenie}{description}
}
\description{
Estimation of Genitor x Progeny Regression
}
\author{
Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva
}
