% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indqi.R
\name{rend_ind}
\alias{rend_ind}
\title{Índice de Descasque e Rendimento Industrial}
\usage{
rend_ind(GEN, NG2M, MG, MC, RG, stat = "all", ...)
}
\arguments{
\item{GEN}{A coluna com o nome dos genótipos.}

\item{NG2M}{Coluna com os valores do número de grãos maior que 2mm.}

\item{MG}{Coluna com os valores da massa de grãos.}

\item{MC}{Coluna com os valores da massa de cariópse.}

\item{RG}{Colunas com os valores do rendimento de grãos (kg/ha).}

\item{stat}{Extrair ou não a média por genótipo. Padrão é `"all"`.}
}
\description{
Cálculo do Índice de Descasque e Rendimento Industrial de Aveia Branca
}
\examples{
\donttest{
library(Breeding)

data("aveia")
# Calcular o rendimento industrial sem extrair a média
with(aveia, rend_ind(GEN, NG2M, MG, MC, RG))

# Calcular o rendimento industrial extraindo a média por genótipo
with(aveia, rend_ind(GEN, NG2M, MG, MC, RG, stat = "mean"))
}
}
