% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restr.R
\name{restr}
\alias{restr}
\title{Restrição da variabilidade das testemunhas}
\usage{
restr(GEN, REP, Xi, scenario = "original")
}
\arguments{
\item{GEN}{A coluna com o nome do genótipo}

\item{REP}{A coluna com as repetições}

\item{Xi}{A coluna com o valor observado para determinado genótipo.}

\item{scenario}{Cenário a ser utilizado para o cálculo.}
}
\value{
Retorna uma tabela com os genótipos e os índices selecionados.
Quanto maior o valor do índice, mais resiliente é o genótipo.
}
\description{
Método para restrição da variabilidade das testemunhas proposta por
Carvalho et al. (2023).
}
\references{
Carvalho, I. R., Silva, J. A. G. da, Moura, N. B., Ferreira, L. L.,
Lautenchleger, F., & Souza, V. Q. de. (2023). Methods for estimation of
genetic parameters in soybeans: An alternative to adjust residual variability.
Acta Scientiarum. Agronomy, 45, e56156.
https://doi.org/10.4025/actasciagron.v45i1.56156
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
