% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risco.R
\name{risco}
\alias{risco}
\title{Estimativa do Risco da Ocorrência de Doenças na Soja}
\usage{
risco(DIA, MES, TEMP, UR, doença = "ferrugem", plot = F)
}
\arguments{
\item{DIA}{Coluna referente ao dia do mês.}

\item{MES}{Coluna referente ao mês do ano (valor numérico).}

\item{TEMP}{Coluna da temperatura média do ar (em ºC).}

\item{UR}{Coluna da umidade relativa (em \%).}

\item{doença}{Definir a doença da soja (Padrão = "ferrugem").}

\item{plot}{Plotar um gráfico do acúmulo (Padrão é F (FALSE)).}
}
\description{
Cálculo do Risco da Ocorrência de Doenças na Soja em função de variáveis
meteorológicas (Engers et al., 2024).
}
\examples{
\donttest{
library(Breeding)

# Predição de Risco de Ferrugem
data("clima")
with(clima, risco(DY, MO, TMED, RH, doença = "ferrugem"))
}
}
\references{
de Oliveira Engers, L.B., Radons, S.Z., Henck, A.U. et al.
Evaluation of a forecasting system to facilitate decision-making for the
chemical control of Asian soybean rust. Trop. plant pathol. 49, 539–546 (2024).
https://doi.org/10.1007/s40858-024-00649-1
}
\author{
Willyan Jr. A. Bandeira, Ivan Ricardo Carvalho
}
