% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{SEGREGAÇÃO_PADRÃO}
\alias{SEGREGAÇÃO_PADRÃO}
\title{Standard Segregation}
\usage{
SEGREGAÇÃO_PADRÃO(MELHORAMENTO)
}
\arguments{
\item{MELHORAMENTO}{Base parameter for the print table function}
}
\description{
Didactic table of standard segregation by generation
}
\author{
Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva
}
