% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiental.R
\name{somatermica}
\alias{somatermica}
\title{Cálculo da Soma Térmica}
\usage{
somatermica(TMED, MONTH, cultura = "milho", plot = T)
}
\arguments{
\item{TMED}{Coluna da temperatura média}

\item{cultura}{Definir a cultura (Padrão = "milho")}

\item{plot}{Plotar um gráfico do acúmulo (Padrão é T (TRUE))}
}
\description{
Cálculo da Soma Térmica de Culturas
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho, Murilo V. Loro,
Leonardo C. Pradebon, José A. G. da Silva
}
