% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamef.R
\name{tamef}
\alias{tamef}
\title{Tamanho Efetivo Populacional}
\usage{
tamef(GEN, SI, NE)
}
\arguments{
\item{GEN}{A coluna com o nome do genótipo (progênie).}

\item{SI}{A coluna com o número de indivíduos selecionados.}

\item{NE}{Número de indivíduos conduzidos no período de seleção.}
}
\description{
Estimativa do tamanho efetivo populacional adaptado por Morais (1997).
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho, Murilo V. Loro,
Leonardo C. Pradebon, José A. G. da Silva
}
