% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{termaldata}
\alias{termaldata}
\title{Data Set with air temperature and incident radiation.}
\format{
A data.frame with 100 observations and 4 variables:
\describe{
  \item{Day}{Column with cycle lenght.}
  \item{Period}{Column with two periods (vegetative and reproductive).}
  \item{Temperature}{Average air temperature values.}
  \item{Radiation}{Incident radiation values.}
}
}
\source{
Simulated data for use.
}
\usage{
termaldata
}
\description{
Data Set with air temperature and incident radiation.
}
\keyword{datasets}
