% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtgerm.R
\name{tindex}
\alias{tindex}
\title{AIC and black oat seed multitrait}
\usage{
tindex(fc, germ, sdm, sl, radl, index = "PI")
}
\arguments{
\item{fc}{First germination count.}

\item{germ}{Germination percentage.}

\item{sdm}{Shoot dry mass.}

\item{sl}{Shoot lenght.}

\item{radl}{Radicle lenght.}

\item{index}{Argument to choose the desired index.}
}
\description{
Multitrait selection index for variables of interest
}
\references{
Moura, N. B., Carvalho, I. R., Silva, J. A. G., Loro, M. V., Barbosa, M. H.,
Lautenchleger, F., Marchioro, V. S., & Souza, V. Q. (2021). Akaike criteria
and selection of physiological multi-character indexes for the production of
black oat seeds. Current Plant Studies, 11, 1-8.
https://doi.org/10.26814/cps2021003
}
\author{
Willyan Junior Adorian Bandeira

Ivan Ricardo Carvalo

Murilo Vieira Loro

Leonardo Cesar Pradebon

Jose Antonio Gonzalez da Silva
}
