% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{transgressivos}
\alias{transgressivos}
\title{Selection by Selection Differential (Mean and Deviations)}
\usage{
transgressivos(Gen, Var, Control, ylab = "Selection", xlab = "Genotypes")
}
\arguments{
\item{Gen}{The column with the genotype name}

\item{Var}{The column with the variable of interest}

\item{Control}{The column with the value of the variable 'X' for the witnesses}

\item{ylab}{The name of the Y axis.}

\item{xlab}{The name of the X axis.}
}
\description{
Selection of Transgressive Genotypes - Selection Differential (SD)
}
\author{
Willyan Junior Adorian Bandeira

Ivan Ricardo Carvalo

Murilo Vieira Loro

Leonardo Cesar Pradebon

Jose Antonio Gonzalez da Silva
}
