% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigo.R
\docType{data}
\name{trigo}
\alias{trigo}
\title{Conjunto de Dados de Trigo}
\format{
Um data.frame com 19 observações e 6 variáveis:
\describe{
  \item{TEST}{Identificaçã do tratamento.}
  \item{CE}{Comprimento da espiga.}
  \item{ME}{Massa da espiga.}
  \item{NGE}{Númoero de grãos da espiga.}
  \item{MGE}{Massa de grãos da espiga.}
  \item{NEE}{Número de espiguetas por espiga}
}
}
\source{
Dados reais de campo para utilização.
}
\usage{
trigo
}
\description{
Conjunto de dados de um experimento de trigo com diferentes manejos de herbicidas.
}
\keyword{datasets}
