% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafarea.R, R/vig.R
\docType{data}
\name{vig}
\alias{vig}
\title{Data Set for Leaf Area Index}
\format{
A data.frame with 10 observations and 5 variables:
\describe{
  \item{GEN}{Column with the genotypes.}
  \item{C}{Leaf lenght}
  \item{L}{Leaf width}
  \item{TNL}{Total number of leaves.}
  \item{TDL}{Total dry leavesh.}
}

A data.frame with 54 observations and 6 variables:
\describe{
  \item{Trat}{Column with treatments.}
  \item{PC}{First Count}
  \item{G}{Germination percentage.}
  \item{CPA}{Length of aerial part.}
  \item{RAD}{Root length.}
  \item{MS}{Seedling dry mass.}
  \item{EC}{See what EC is.}
}
}
\source{
Simulated data.

Real field data for use.
}
\usage{
vig

vig
}
\description{
Data set with 10 genotypes and values for leaf length, leaf width, number
of total leaves and number of dry leaves

Data set from experiment with wheat genotypes subjected to different
sowing density.
}
\keyword{datasets}
