% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{ALELICA}
\alias{ALELICA}
\title{Interações alélicas}
\usage{
ALELICA(type = NULL, ge = NULL)
}
\arguments{
\item{type}{Tipo de interação alélica. Utilize "ad" para aditividade, "dom"
para dominância completa, "domp" para dominância parcial e "sob" para
sobredominância.}

\item{ge}{Tipo de interação GxE. Utilize "aus" para ausência de interação,
"simples" para interação simples e "complex" para interação complexa.}
}
\description{
Exemplos didáticos de interações alélicas e gênicas
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
