% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{ALELICA}
\alias{ALELICA}
\title{Allelic interactions}
\usage{
ALELICA(type = NULL, ge = NULL)
}
\arguments{
\item{type}{Type of allelic interaction. Use “ad” for additivity, “dom”
for complete dominance, “domp” for partial dominance and “sob” for
overdominance.}

\item{ge}{Type of GxE interaction. Use “aus” for no interaction,
“simple” for simple interaction and “complex” for complex interaction.}
}
\description{
Examples of allelic and gene interactions
}
\author{
Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva
}
