% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{blupis}
\alias{blupis}
\title{Number of individuals to be selected in each family}
\usage{
blupis()
}
\arguments{
\item{y}{Response variable}

\item{r}{The column with the repetition effect (fixed)}

\item{p}{The column with the plot effect (random)}

\item{f}{The column with the genetic effect of dominance associated with
families of full siblings (random)}

\item{b}{The column with the effect of incomplete blocks (random)}

\item{A}{Kinship matrix obtained with AGHmatrix}
}
\description{
Definition of the number of individuals to be selected by BLUPis
}
\references{
Oliveira, R. A., Daros, E., Resende, M. D. V., Bespalhok-Filho, J. C.,
Zambon, J. L. C., Souza, T. R., & Lucius, A. S. F. (2011). Procedimento
Blupis e seleção massal em cana-de-açúcar. Bragantia, 70(4), 796–800.
Model 35 (SELEGEN)
}
\author{
Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva
}
