% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{clima}
\alias{clima}
\title{Data: Climate Data Set for Predictions}
\format{
A data.frame with 8760 observations and 5 variables:
\describe{
  \item{MO}{Month of the year.}
  \item{DY}{Day of the year.}
  \item{HR}{Time of the day.}
  \item{TMED}{Average Air Temperature - in  degree C.}
  \item{RH}{Relative Humidity - in \%.}
}
}
\source{
Data obtained from the Nasa Power platform (https://power.larc.nasa.gov/).
}
\usage{
clima
}
\description{
Average air temperature and relative humidity data for the period of
one year, with time, day and month.
}
\keyword{datasets}
