% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{coefend}
\alias{coefend}
\title{Data: Data: Endogamy Coefficient Data Set}
\format{
A data.frame with 7 observations and 5 variables:
\describe{
  \item{Var}{Variable name.}
  \item{VF}{Phenotypic Variance.}
  \item{VG}{Genotypic Variance.}
  \item{h}{Broad-sense heritability}
  \item{DS}{Selection Differential}
}
}
\source{
Real data for use.
}
\usage{
coefend
}
\description{
Data set of phenotypic and genotypic variance, heritability and
differential selection for different variables.
}
\keyword{datasets}
