% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{Coeficiente_endogamia}
\alias{Coeficiente_endogamia}
\title{Inbreeding coefficient}
\usage{
Coeficiente_endogamia(var, VG, VF)
}
\arguments{
\item{var}{Column with the variable name}

\item{VG}{Column with genotypic variance}

\item{VF}{Column with phenotypic variance}
}
\description{
Function for calculating the inbreeding coefficient
}
\author{
Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva
}
