% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{Coeficiente_endogamia}
\alias{Coeficiente_endogamia}
\title{Função para Cálculo do coeficiente de endogamia}
\usage{
Coeficiente_endogamia(var, VG, VF)
}
\arguments{
\item{var}{description}

\item{VG}{description}

\item{VF}{description}
}
\description{
Função para Cálculo do coeficiente de endogamia
}
