% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{Coeficiente_endogamia}
\alias{Coeficiente_endogamia}
\title{Coeficiente de endogamia}
\usage{
Coeficiente_endogamia(var, VG, VF)
}
\arguments{
\item{var}{Coluna com o nome da variável}

\item{VG}{Coluna com a variância genotípica}

\item{VF}{Coluna com a variância fenotípica}
}
\description{
Função para Cálculo do coeficiente de endogamia
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
