% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restr.R
\name{cvar}
\alias{cvar}
\title{Estimation of variance components by restricting the variability of the
witnesses.}
\usage{
cvar(GEN, REP, Xi, approach = NULL, zstat = NULL)
}
\arguments{
\item{GEN}{The column with the name of the genotypes (without controls).}

\item{REP}{The column with the repetitions (if any).}

\item{Xi}{The column with the observed value for the variable in a given genotype.}

\item{approach}{Method to be used for estimating variance components. Use “apI”
for parent-offspring regression, “apII” for the of the sum of squares of
augmented blocks with intercalary parents, “apIII” for the method with linear
 mixed models with random genetic effects.}

\item{zstat}{Logical argument. Applies Z-notation normalization if “TRUE”.}
}
\description{
Estimation of variance components and genetic parameters from the restriction
of witness values
}
\references{
Carvalho, I. R., Silva, J. A. G. da, Moura, N. B., Ferreira, L. L.,
Lautenchleger, F., & Souza, V. Q. de. (2023). Methods for estimation of
genetic parameters in soybeans: An alternative to adjust residual variability.
Acta Scientiarum. Agronomy, 45, e56156.
https://doi.org/10.4025/actasciagron.v45i1.56156
}
\author{
Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva
}
