% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{default_seg}
\alias{default_seg}
\title{Standard Segregation}
\usage{
default_seg(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical argument. Runs the code silently if FALSE.}
}
\value{
Create a didactic table of standard segregation, considering allogamous
and autogamous species and mutants. It shows the expected level of
heterozygosity, probable number of genes, environmental effect and Wright's
probabilistic coefficient.
}
\description{
Didactic table of standard segregation by generation
}
\examples{
library(EstimateBreed)

default_seg(verbose=TRUE)
}
\author{
Willyan Junior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

Jose Antonio Gonzalez da Silva
}
