% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiental.R
\name{deltat}
\alias{deltat}
\title{Aplicação de defensivos agrícolas}
\usage{
deltat(
  LON,
  LAT,
  type = 2,
  days = 7,
  control = NULL,
  details = FALSE,
  dates = NULL,
  plot = FALSE
)
}
\arguments{
\item{LON}{Longitude (em decimal)}

\item{LAT}{Latitude (em decimal)}

\item{type}{Tipo de análise. Utilize 1 para forecast e 2 para dados temporais.}

\item{days}{Número de dias}

\item{control}{Tipo de produto a ser aplicado. Utilizar "fung" para fungicida,
"herb" para herbicida, "ins" para inseticidas, "bio" para produtos biológicos.}

\item{details}{Retorna o resultado de forma detalhada se TRUE.}

\item{dates}{Só utilizar esse argumento se type=2. Data de início e final
para a obtenção dos dados meteorológicos para um ciclo de cultivo.}
}
\value{
Retorna os momentos ideais de aplicação, considerando cada cenário.
Tomado como parâmetro um DELTA_T entre 2 e 8, velocidade do vento entre 3 e 8,
e ausência de precipitação.
}
\description{
Determinação do momento ideal para aplicação
}
\examples{
\donttest{
library(Breeding)

# Previsão das condições de aplicação
deltat(-53.696944444444,-28.063888888889,type=1,days=10)
View(forecast)

# Análise retrospectiva das condições de aplicação
deltat(-53.696944444444,-28.063888888889,type=1,days=10,
dates=c("2023-01-01","2023-05-01"))
View(retrospective)
}
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho, Murilo V. Loro,
Leonardo C. Pradebon, José A. G. da Silva
}
