% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isgr.R
\name{desv_clim}
\alias{desv_clim}
\title{Função auxiliar para o cálculo do ISGR}
\usage{
desv_clim(ENV, TMED, PREC)
}
\arguments{
\item{ENV}{Identificação de cada ambiente de seleção (para diferenciar caso
haja mais de um ciclo de cultivo).}

\item{TMED}{Temperatura média do ar (em ºC) ocorrida durante o ciclo de
cultivo em cada ambiente.}

\item{PREC}{Precipitação pluviométrica (em mm) ocorrida durante o ciclo de
cultivo em cada ambiente.}
}
\value{
Um dataframe contendo o identificador do ambiente de seleção e
os desvios padrões para temperatura e precipitação.
}
\description{
Esta função recebe um dataframe com os dados de temperatura e precipitação
e calcula o desvio padrão desses parâmetros para cada ambiente.
}
\examples{
\donttest{
library(Breeding)
data("desvamb")

with(desvamb,desv_clim(ENV,TMED,PREC))
}
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
