% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{desvamb}
\alias{desvamb}
\title{Data: Data set for calculating the environmental deviation}
\format{
A data.frame with 449 observations and 3 variables:
\describe{
  \item{ENV}{Selection environment.}
  \item{TMED}{Average Air Temperature (in degree C).}
  \item{PREC}{Precipitation (in mm)}
}
}
\source{
Real field data for use.
}
\usage{
desvamb
}
\description{
Data set with average air temperature and precipitation values per environment
}
\keyword{datasets}
