% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defh.R
\name{dhid}
\alias{dhid}
\title{Determinação do Déficit Hídrico}
\usage{
dhid(CICLO, FASE, ETo, AD = 4, PROF = 100, Cultura = "Soja")
}
\arguments{
\item{CICLO}{Coluna referente ao ciclo da cultura}

\item{FASE}{Fase em que a cultura se encontra}

\item{ETo}{Evapotranspiração potencial do dia}

\item{PROF}{Profundidade Aproximada do Solo (em cm)}

\item{Cultura}{Definir a cultura (Padrão = "Soja")}

\item{ADP}{Água Disponível Predita}

\item{plot}{Plotar um gráfico do acúmulo (Padrão é F (FALSE))}
}
\description{
Estimativa do déficit hídrico das culturas agrícolas
}
\references{
Bortoluzzi, M. P., Heldwein, A. B., Trentin, R., Maldaner, I. C., & da Silva,
J. R. (2020). Risk of occurrence of water deficit in soybean cultivated in
lowland soils. Earth Interactions, 24(4), 1-16.
https://doi.org/10.1175/EI-D-19-0029.1
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho, Murilo V. Loro,
Leonardo C. Pradebon, José A. G. da Silva
}
