% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estresse.R
\name{estresse}
\alias{estresse}
\title{Stress indices for genotype selection}
\usage{
estresse(
  GEN,
  YS,
  YC,
  index = "ALL",
  bygen = T,
  plot = F,
  xlab = "Genotype",
  ylab = "Values",
  ...
)
}
\arguments{
\item{GEN}{The column with the genotypes to be selected.}

\item{YS}{Productivity of the genotype without stress conditions.}

\item{YC}{Genotype productivity under stressful conditions.}

\item{index}{Index to be calculated (Standard “ALL”). The indices to be used
are: 'STI' - Stress Tolerance Index, 'YI' - Yield Index, 'GMP' - Geometric Mean
Productivity, 'MP' - Mean Productivity, 'MH' - Harmonic Mean, 'SSI' - Stress
Stability Index, 'YSI' - Yield Stability Index, 'RSI' - Relative Stress Index.}

\item{bygen}{Returns the average of each genotype if “TRUE”. Only in this way
it will be possible to plot graphs.}

\item{plot}{Plot graph if equal to “TRUE” (Standard “F”).}

\item{xlab}{Adjust the title of the x-axis in the graph.}

\item{ylab}{Adjust the title of the y-axis in the graph.}
}
\value{
Returns a table with the genotypes and the selected indices.
The higher the index value, the more resilient the genotype.
}
\description{
Selection indices for genotypes conducted under stress conditions cited
by Ghazvini et al. (2024).
}
\examples{
\donttest{
library(Breeding)

data("aveia")
with(aveia,estresse(GEN,MC,MG,index = "ALL",bygen=T))

}
}
\references{
Ghazvini, H., Pour-Aboughadareh, A., Jasemi, S.S. et al.
A Framework for Selection of High-Yielding and Drought-tolerant
Genotypes of Barley: Applying Yield-Based Indices and Multi-index
Selection Models. Journal of Crop Health 76, 601–616 (2024).
https://doi.org/10.1007/s10343-024-00981-1
}
\author{
Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva
}
