% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiental.R
\name{fototermal}
\alias{fototermal}
\title{Photothermal Index}
\usage{
fototermal(DAY, TMED, RAD, PER)
}
\arguments{
\item{DAY}{The column with the cycle days}

\item{TMED}{The column with the average air temperature values}

\item{RAD}{The column with the incident radiation values}

\item{PER}{The column with the period (use VEG for vegetative and REP for
reproductive)}
}
\description{
Calculation of the photothermal index based on average temperature and
radiation
}
\references{
Zanon, A. J., & Tagliapietra, E. L. (2022). Ecofisiologia da soja:
Visando altas produtividades (2a ed.). Field Crops.
}
\author{
Willyan Junior Adorian Bandeira

Ivan Ricardo Carvalo

Murilo Vieira Loro

Leonardo Cesar Pradebon

Jose Antonio Gonzalez da Silva
}
