% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{genot}
\alias{genot}
\title{Data: GxE Interaction}
\format{
A data.frame with 55 observations and 5 variables:
\describe{
  \item{GEN}{Selected lines in a GXE experiment.}
  \item{ENV}{Selection environments.}
  \item{NG}{Number of grains measured in the lines.}
  \item{MG}{Grain mass measured in the lines (in g)}
  \item{CICLO}{Length of crop cycle (in days)}
}
}
\source{
Real field data for use.
}
\usage{
genot
}
\description{
Data set with strains and test subjects from a GxE experiment.
}
\keyword{datasets}
