% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{genpar}
\alias{genpar}
\title{General parameters for selection}
\usage{
genpar(POP, GEN, REP = NULL, vars, K = 0.05, type = "balanced", check = FALSE)
}
\arguments{
\item{POP}{The column with the population under improvement.}

\item{GEN}{The column with the selected genotypes within the population.}

\item{REP}{The column with the repetitions (if any).}

\item{K}{Selection pressure (Default 0.05).}

\item{type}{Type of experiment (balanced or unbalanced). Use
“balanced” for balanced and “unb” for unbalanced.}

\item{check}{Logical argument. Checks the model's assumptions
statistical if the value is equal to TRUE.}

\item{VAR}{The column with the variable of interest.}
}
\description{
Function for determining selection parameters, based on an experiment
carried out on the rice crop. Intended for isolated evaluation of the performance
of strains within a given population.
}
\references{
Yadav, S. P. S., Bhandari, S., Ghimire, N. P., Mehata, D. K., Majhi, S. K.,
Bhattarai, S., Shrestha, S., Yadav, B., Chaudhary, P., & Bhujel, S. (2024).
Genetic variability, character association, path coefficient, and diversity
analysis of rice (Oryza sativa L.) genotypes based on agro-morphological
traits. International Journal of Agronomy, 2024, Article ID 9946332.
https://doi.org/10.1155/2024/9946332
}
\author{
Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva
}
