% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{gga}
\alias{gga}
\title{Additive Genetic Gain}
\usage{
gga(GEN, VAR, h2, P)
}
\arguments{
\item{GEN}{Vector or dataframe containing the genotypes to be selected.}

\item{VAR}{Variable of interest for analysis.}

\item{h2}{Heritability of the character (a value between 0 and 1).}

\item{P}{Performance of the progeny (a numerical value or vector).}
}
\description{
Estimates the additive genetic gain, as described by Falconer (1987).
}
\author{
Willyan Junior Adorian Bandeira

Ivan Ricardo Carvalo

Murilo Vieira Loro

Leonardo Cesar Pradebon

Jose Antonio Gonzalez da Silva
}
