% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{gga}
\alias{gga}
\title{Ganho genético aditivo}
\usage{
gga()
}
\arguments{
\item{GEN}{Genótipos a serem selecionados.}

\item{VAR}{Variável de interesse.}

\item{h2}{Valor da herdabilidade.}

\item{P}{Performance da progênie.}

\item{u}{Média do grupo contemporâneo mais 10%.}
}
\description{
Estimativa do ganho genético aditivo (Falconer, 1987).
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
