% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{gga}
\alias{gga}
\title{Ganho Genético Aditivo}
\usage{
gga(GEN, VAR, h2, P)
}
\arguments{
\item{GEN}{Vetor ou dataframe contendo os genótipos a serem selecionados.}

\item{VAR}{Variável de interesse para análise.}

\item{h2}{Herdabilidade do caráter (um valor entre 0 e 1).}

\item{P}{Performance da progênie (um valor numérico ou vetor).}
}
\description{
Esta função estima o ganho genético aditivo, conforme descrito por Falconer (1987).
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
