% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{GS}
\alias{GS}
\title{Pressão de Seleção}
\usage{
GS(Var, h, VF, P = "1")
}
\arguments{
\item{Var}{A coluna com o nome das variáveis de interesse}

\item{h}{A coluna com os valores de herdabilidade no sentido restrito}

\item{VF}{A coluna com os valores de variância fenotípica}

\item{P}{A coluna com os valores observados para as progênies}
}
\description{
Resposta a seleção ponderada pela pressão de seleção
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
