% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{GS}
\alias{GS}
\title{Resposta a seleção ponderada pela pressão de seleção}
\usage{
GS(Var, h, VF, P = "1")
}
\arguments{
\item{Var}{Nome da variável resposta}

\item{h}{Nome da coluna com o valor da herdabilidade}

\item{VF}{Nome da coluna com o valor da variância fenotípica}

\item{P}{Valor da pressão de seleção (padrão = 1)}
}
\description{
Resposta a seleção ponderada pela pressão de seleção
}
