% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{GS2}
\alias{GS2}
\title{Resposta a seleção ponderada pelo Diferencial de Seleção}
\usage{
GS2(Var, h, DS)
}
\arguments{
\item{Var}{description}

\item{h}{description}

\item{DS}{description}
}
\description{
Resposta a seleção ponderada pelo Diferencial de Seleção
}
