% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{GS2}
\alias{GS2}
\title{Single Selection Differential}
\usage{
GS2(Var, h, DS)
}
\arguments{
\item{Var}{The column with the name of the variables of interest}

\item{h}{The column with the heritability values in the strict sense}

\item{DS}{The column with the value of the selection differential to be applied
to each variable}
}
\description{
Response to selection weighted by the Simple Selection Differential
}
\author{
Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva
}
