% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seleção.R
\name{GS3}
\alias{GS3}
\title{Resposta a seleção ponderada pela pressão de seleção e Controle de Genitores}
\usage{
GS3(Var, h, VF, P = "1")
}
\arguments{
\item{Var}{description}

\item{h}{description}

\item{VF}{description}

\item{P}{description}
}
\description{
Considera conhecer apenas o genitor materno, sem controle do polinizador,
ou uma seleção direta sem parentais
}
