% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{GS3}
\alias{GS3}
\title{Resposta a Seleção pelo Controle de Genitores}
\usage{
GS3(Var, h, VF, P = "1")
}
\arguments{
\item{Var}{A coluna com as variáveis de interesse}

\item{h}{A coluna com os valores de herdabilidade no sentido restrito}

\item{VF}{A coluna com os valores de variância fenotípica}

\item{P}{A coluna com os valores das progênies}
}
\description{
Considera conhecer apenas o genitor materno, sem controle do polinizador,
ou uma seleção direta sem parentais
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho, Murilo V. Loro,
Leonardo C. Pradebon, José A. G. da Silva
}
