% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seleção.R
\name{GS4}
\alias{GS4}
\title{Resposta a seleção ponderada pela Pressão Ponderado pelo Anode seleção e ano}
\usage{
GS4(Var, h, VF, P = "1", Ano)
}
\arguments{
\item{Var}{description}

\item{h}{description}

\item{VF}{description}

\item{P}{description}

\item{Ano}{description}
}
\description{
Resposta a seleção ponderada pela Pressão Ponderado pelo Anode seleção e ano
}
