% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{GS4}
\alias{GS4}
\title{Resposta Seleção pelo Ano}
\usage{
GS4(Var, h, VF, P = "1", Ano)
}
\arguments{
\item{Var}{A coluna com a variáveis de interesse}

\item{h}{A coluna com os valores de herdabilidade no sentido restrito}

\item{VF}{A coluna com os valores de variância fenotípica}

\item{P}{A coluna com o valor obtido para as progênies}

\item{Ano}{A coluna com o ano de seleção}
}
\description{
Resposta a seleção ponderada pela Pressão Ponderado pelo Ano de seleção e Ano
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
