% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvri.R, R/indgerm.R
\name{gvri}
\alias{gvri}
\title{Índice de Seleção para Volume de Grão}
\usage{
gvri(GEN, C, L, E, stat = "all", plot = F, ylab = "GVRI", xlab = "Genótipo")

gvri(GEN, C, L, E, stat = "all", plot = F, ylab = "GVRI", xlab = "Genótipo")
}
\arguments{
\item{GEN}{A coluna com o nome dos genótipos.}

\item{C}{Comprimento do Grãos}

\item{L}{Largura do Grão}

\item{E}{Espessura do Grão}

\item{stat}{Extrair ou não a média por genótipo. Padrão é `"all"`.}

\item{plot}{Plotar o gráfico se `T`}

\item{ylab}{Nome do eixo Y}

\item{xlab}{Nome do eixo X}
}
\description{
Cálculo do Índice de Seleção para Volume de Grão

Índice de seleção para volume de grão proposto por Carvalho et al. (2017).
}
\references{
Carvalho, I. R., Pelegrin, A. D., Szareski, V. J., Ferrari, M.,
Rosa, T. D., Martins, T. S., ... & Maia, L. D. (2017).
Diallel and prediction (REML/BLUP) for yield components in intervarietal maize hybrids.
Genet Mol Res, 16(3).
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho, Murilo V. Loro,
Leonardo C. Pradebon, José A. G. da Silva
}
