% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterose.R
\name{heterose}
\alias{heterose}
\title{Função para cálculo da heterose e heterobeltiose}
\usage{
heterose(GEN, GM, GP, PR, param = "all")
}
\arguments{
\item{GEN}{A coluna com o nome do genótipo}

\item{GM}{A coluna com a média do genitor materno}

\item{GP}{A coluna com a média do genitor paterno}

\item{PR}{A coluna com a média da progênie}

\item{param}{Valor para determinar o parâmetro a ser calculado. Padrão é "all".
Para calcular apenas a heterose, utilize "het". Para calcular apenas a
heterobeltiose, utilize "hetb".}
}
\description{
Cálculo dos parâmetros da heterose e heterobeltiose de híbridos de milho
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho.
}
