% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterose.R
\name{heterose}
\alias{heterose}
\title{Heterosis and Heterobeltiosis}
\usage{
heterose(GEN, GM, GP, PR, REP, param = "all")
}
\arguments{
\item{GEN}{The column with the genotype name}

\item{GM}{The column with the average of the maternal parent}

\item{GP}{The column with the average of the paternal parent}

\item{PR}{The column with the average of the progeny}

\item{REP}{The column with the repetitions}

\item{param}{Value to determine the parameter to be calculated. Default is “all”.
To calculate heterosis only, use “het”. To calculate only heterobeltiosis,
use “hetb”.}
}
\description{
Calculation of heterosis and heterobeltiosis parameters of hybrids
}
\examples{
\donttest{
library(Breeding)

data("maize")
#Extract heterosis and heterobeltiosis
with(maize,heterose(HIB,GM,GP,P,stat="all"))

#Only extract heterosis
with(maize,heterose(HIB,GM,GP,P,param = "het"))

#Extract only heterobeltiosis
with(maize,heterose(HIB,GM,GP,P,param = "hetb"))
}
}
\author{
Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva
}
