% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaf.R
\name{iaf}
\alias{iaf}
\title{Estimativa do Índice de Área Foliar (IAF)}
\usage{
iaf(GEN, W, L, crop = "soy", sp = 0.45, sden = 14)
}
\arguments{
\item{GEN}{Colune com o nome do genótipo}

\item{W}{Largura da Folha.}

\item{L}{Comprimento da Folha.}

\item{crop}{Cultura amostrada. Utilizar "soy" para soja e "maize" para milho.}

\item{sp}{Espaçamento entre linhas (Padrão sp=0.45).}

\item{sden}{Densidade de semeadura, em plantas por metro linear (padrão sden=14).}
}
\description{
Função utilitária para a determinação do IAF de culturas
}
\examples{
\donttest{
library(EstimateBreed)

#Finalizar o exemplo
}
}
\references{
Meira, D., Queiróz de Souza, V., Carvalho, I. R., Nardino, M., Follmann,
D. N., Meier, C., Brezolin, P., Ferrari, M., & Pelegrin, A. J. (2015).
Plastocrono e caracteres morfológicos da soja com hábito de crescimento
indeterminado. Revista Cultivando o Saber, 8(2), 184-200.
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
