% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indgerm.R
\name{index_vigor}
\alias{index_vigor}
\title{Complete vigor index}
\usage{
index_vigor(GEN, PC, G, CPA, RAD, MS, EC)
}
\arguments{
\item{GEN}{The column with the genotype name}

\item{PC}{The column with the values from the first count}

\item{G}{The column with the values of percentage of germinated seeds}

\item{CPA}{The column with the values of shoot length}

\item{RAD}{The column with the values of Root Length}

\item{MS}{The column with the values of dry mass}

\item{EC}{TThe column with the field emergency values}
}
\description{
Determining the multivariate vigor of seeds
}
\references{
Szareski, V. J., Carvalho, I. R., Demari, G. H., Rosa, T. C. D.,
Souza, V. Q. D., Villela, F. A.,Aumonde, T. Z. (2018).
Multivariate index of soybean seed vigor:
a new biometric approach applied to the effects of genotypes and
environments. Journal of Seed Science, 40(4), 396-406.
}
\seealso{
\code{\link{ivig_simp}}
}
\author{
Willyan Junior Adorian Bandeira

Ivan Ricardo Carvalo

Murilo Vieira Loro

Leonardo Cesar Pradebon

Jose Antonio Gonzalez da Silva
}
