% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indgerm.R
\name{indger}
\alias{indger}
\title{Germination index by subsequent counting}
\usage{
indger(TESTE, DIA, TSG, NST)
}
\arguments{
\item{TESTE}{Identification number of the test performed}

\item{DIA}{Numerical values for the days tested}

\item{TSG}{Name of the column with the total number of seeds germinated in
each period}

\item{NST}{Default column name with number of seeds tested}
}
\value{
Returns the germination index by subsequent count.
}
\description{
Estimation of germination index by subsequent seedling count
germinated in a given period (Wang et al., 2017)
}
\references{
Wang, Pan, Li, Dong, Wang, Li-jun and Adhikari, Benu. "Effect of High
Temperature Intermittent Drying on Rice Seed Viability and Vigor" International
Journal of Food Engineering, vol. 13, no. 10, 2017, pp. 20160433.
https://doi.org/10.1515/ijfe-2016-043
}
\author{
Willyan Junior Adorian Bandeira

Ivan Ricardo Carvalo

Murilo Vieira Loro

Leonardo Cesar Pradebon

Jose Antonio Gonzalez da Silva
}
