% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indqi.R
\name{indviab}
\alias{indviab}
\title{Estimativa de índices de espiga}
\usage{
indviab(
  GEN,
  var1,
  var2,
  ylab = "Índice",
  xlab = "Genótipo",
  stat = "all",
  plot = F
)
}
\arguments{
\item{var1}{Nome da coluna que contém a primeira variável}

\item{var2}{Nome da coluna que contém a segunda variável}

\item{genot}{Nome da coluna que contém os genótipos}
}
\description{
Estimativa do índice de viabilidade a partir de duas variáveis de campo.
}
\examples{
with(data,indviab(genot,var1,var2)) #Ajustar o exemplo###
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
