% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indqi.R
\name{is_ptnerg}
\alias{is_ptnerg}
\title{Índice de seleção para proteína x rendimento de grãos}
\usage{
is_ptnerg(GEN, PTN, RG)
}
\arguments{
\item{GEN}{A coluna com o nome dos genótipos}

\item{PTN}{Proteina Bruta}

\item{RG}{Rendimento de grãos}
}
\description{
Índice de seleção para proteína e rendimento de grãos (Pelegrin et al., 2017).
}
\references{
de Pelegrin, A. J., Carvalho, I. R., Nunes, A. C. P., Demari, G. H., Szareski,
V. J., Barbosa, M. H., ... & da Maia, L. C. (2017).
Adaptability, stability and multivariate selection by mixed models.
American Journal of Plant Sciences, 8(13), 3324.
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
