% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estresse.R
\name{itu}
\alias{itu}
\title{Environmental Stress Index}
\usage{
itu(AAT, RH)
}
\arguments{
\item{AAT}{The column with the average air temperature values}

\item{RH}{The column with the relative humidity values}
}
\value{
Returns the stress condition based on the reported air temperature and
relative humidity values, being: Non-stressful condition (ITU>=70), Heat
stress condition (ITU between 71 and 78), Severe heat stress (ITU between 79
and 83), and Critical heat stress condition (ITU above 84).
}
\description{
Determining the UTI (temperature and humidity index) from the air temperature
and relative humidity values over a given period of time
}
\references{
Tazzo, I. F., Tarouco, A. K., Allem Junior P. H. C., Bremm, C., Cardoso, L.
S., & Junges, A. H. (2024). Indice de Temperatura e Umidade (ITU) ao longo do
verao de 2021/2022 e estimativas dos impactos na bovinocultura de leite no Rio
Grande do Sul, Brasil. Ciencia Animal Brasileira, 2,5, e-77035P.
}
\author{
Willyan Junior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

Jose Antonio Gonzalez da Silva
}
