% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indgerm.R
\name{ivig_simp}
\alias{ivig_simp}
\title{Simple Vigor Index}
\usage{
ivig_simp(GEN, PC, G)
}
\arguments{
\item{GEN}{The column with the genotypes.}

\item{PC}{First count values}

\item{G}{Germination percentage}
}
\description{
Simple seed vigor index described by Szareski et al. (2018).
}
\references{
Szareski, V. J., Carvalho, I. R., Kehl, K., Levien, A. M., Nardino,
M., Dellagostin, S. M., ... & Aumonde, T. Z. (2018).
Adaptability and stability of wheat genotypes according to the
phenotypic index of seed vigor. Pesquisa Agropecuaria Brasileira,
53, 727-735.
}
\seealso{
\code{\link{index_vigor}}
}
\author{
Willyan Junior Adorian Bandeira

Ivan Ricardo Carvalo

Murilo Vieira Loro

Leonardo Cesar Pradebon

Jose Antonio Gonzalez da Silva
}
