% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JPooni.R
\name{Jinks_Pooni}
\alias{Jinks_Pooni}
\title{Função para estimar a probabilidade de extrair linhagens superiores de populações
Método de Jinks and Pooni}
\usage{
Jinks_Pooni(População, Var, VG, Testemunhas)
}
\arguments{
\item{name}{description}
}
\description{
Port, E. D., Carvalho, I. R., Pradebon, L. C., Loro, M. V., Colet, C. D. F.,
Silva, J. A. G. D., & Sausen, N. H. (2024).
Early selection of resilient progenies to seed yield in soybean populations.
 Ciência Rural, 54, e20230287.
}
