% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{Jinks_Pooni}
\alias{Jinks_Pooni}
\title{Jinks and Pooni method}
\usage{
Jinks_Pooni(Pop, Var, VG, Test)
}
\arguments{
\item{Pop}{The column with the population name}

\item{Var}{The column with the variable name}

\item{VG}{The column with the genotypic variance values}

\item{Test}{The column with the controls names}
}
\description{
Function for estimating the probability of extracting superior lines from
populations by the Jinks and Pooni method
}
\references{
Port, E. D., Carvalho, I. R., Pradebon, L. C., Loro, M. V., Colet, C. D. F.,
Silva, J. A. G. D., & Sausen, N. H. (2024).
Early selection of resilient progenies to seed yield in soybean populations.
Ciencia Rural, 54, e20230287.
}
\author{
Willyan Junior Adorian Bandeira

Ivan Ricardo Carvalo

Murilo Vieira Loro

Leonardo Cesar Pradebon

Jose Antonio Gonzalez da Silva
}
