% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{lin}
\alias{lin}
\title{Data: Wheat Data Set with Protein and Grain Yield}
\format{
A data.frame with 24 observations and 7 variables:
\describe{
  \item{POP}{Base population.}
  \item{MGP_MF}{Phenotypic average of grain mass per plant.}
  \item{MGP_GP}{Genotypic average of grain mass per plant.}
  \item{VF}{Phenotypic variance}
  \item{VG}{Genetic variance}
  \item{H2}{Heritability in the broad sense}
  \item{Test}{Witness parameters}
}
}
\source{
Real field data for use.
}
\usage{
lin
}
\description{
Data set with wheat genotypes, protein percentage and grain yield.
}
\keyword{datasets}
