% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearest.R
\name{linearest}
\alias{linearest}
\title{Estimativas a partir de equações polinomiais.}
\usage{
linearest(indep, dep, type = NULL, alpha = 0.05)
}
\arguments{
\item{indep}{Nome da coluna com a variável independente.}

\item{dep}{Nome da coluna da variável dependente}

\item{type}{Tipo de análise a ser realizada.}

\item{alpha}{Significância do teste}
}
\value{
Retorna uma tabela com os genótipos e os índices selecionados.
Quanto maior o valor do índice, mais resiliente é o genótipo.
}
\description{
Determinação da máxima eficiência técnica (MET), x2, x3 e platô.
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
