% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearest.R
\name{linearest}
\alias{linearest}
\title{Estimates using polynomial equations.}
\usage{
linearest(indep, dep, type = NULL, alpha = 0.05)
}
\arguments{
\item{indep}{Name of the column with the independent variable.}

\item{dep}{Name of the dependent variable column}

\item{type}{Type of analysis to be carried out. Use 'MET' to extract the
maximum technical efficiency, 'x3' to obtain the maximum and minimum points
and 'plateau' to extract the parameters using the plateau function.}

\item{alpha}{Significance of the test.}
}
\description{
Determination of maximum technical efficiency (MET), maximum and minimum points
and plateau function.
}
\author{
Willyan Junior Adorian Bandeira

Ivan Ricardo Carvalo

Murilo Vieira Loro

Leonardo Cesar Pradebon

Jose Antonio Gonzalez da Silva
}
