% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{maize}
\alias{maize}
\title{Data: Maize Dataset}
\format{
A data.frame with 4 observations and 3 variables:
\describe{
  \item{P}{Progenies.}
  \item{GM}{Maternal Parent}
  \item{GP}{Patern Parent}
}
}
\source{
Simulated Data.
}
\usage{
maize
}
\description{
Data set with progenies and maternal and paternal maize genitors.
}
\keyword{datasets}
