% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutagenico.R
\name{mut_index}
\alias{mut_index}
\title{Índice de vigor multivariado em sementes
Citação: "A new method for mutation inducing in rice
by using DC electrophoresis bath and its mutagenic ef}
\usage{
mut_index(mut, MSG, MST, GT, DT, SL)
}
\arguments{
\item{mut}{Método de mutação}

\item{MSG}{Média de Sementes Germinadas}

\item{MST}{Média de Sementes Total}

\item{GT}{Número de plântulas por dia durante o tempo t}

\item{DT}{Número de Dias Avaliados}

\item{SL}{Comprimento da plântula}
}
\description{
Índice de vigor multivariado em sementes
Citação: "A new method for mutation inducing in rice
by using DC electrophoresis bath and its mutagenic ef
}
