% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutagenico.R
\name{mut_index}
\alias{mut_index}
\title{Multivariate seed vigor index}
\usage{
mut_index(mut = NULL, MSG, MST, GT, DT, SL)
}
\arguments{
\item{mut}{Mutation method}

\item{MSG}{The column with the average number of germinated seeds}

\item{MST}{The column with the average total seeds}

\item{GT}{Number of seedlings germinated per day during 't' time}

\item{DT}{Number of evaluation days}

\item{SL}{Shoot Length}
}
\description{
Determining the vigor of seeds obtained from mutation induction processes
}
\references{
Zou, M., Tong, S., Zou, T. et al. A new method for mutation inducing in rice
by using DC electrophoresis bath and its mutagenic effects. Sci Rep 13, 6707
(2023). https://doi.org/10.1038/s41598-023-33742-7
}
\author{
Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva
}
