% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiental.R
\name{optemp}
\alias{optemp}
\title{Plotting the optimum and cardinal temperatures for crops}
\usage{
optemp(
  DAS,
  Var,
  crop = "soybean",
  ylab = "Meteorological Atribute",
  xlab = "Days After Sowing"
)
}
\arguments{
\item{DAS}{Days after sowing}

\item{Var}{desc}

\item{crop}{Soja, Milho, Trigo}

\item{ylab}{desc}

\item{xlab}{description}
}
\description{
Utility function for plotting graphs of thermal preferences for crops
}
\examples{
\donttest{
library(EstimateBreed)

}
}
\author{
Willyan Junior Adorian Bandeira

Ivan Ricardo Carvalo

Murilo Vieira Loro

Leonardo Cesar Pradebon

Jose Antonio Gonzalez da Silva
}
