% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indqi.R
\name{ph}
\alias{ph}
\title{Determinação do peso do hectolitro de cereais}
\usage{
ph(GEN, PESO, crop = "trigo")
}
\arguments{
\item{GEN}{A coluna com o nome do genótipo}

\item{PESO}{Peso obtido em balança de 1qt lt}
}
\value{
Retorna o valor estimado do peso do hectolitro (PH).
}
\description{
Função útil para caracterizar o peso do hectolitro (PH) de experimentos com
cereais.
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
