% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{pheno}
\alias{pheno}
\title{Soybean Plastochron Estimation Data Set}
\format{
A data.frame with 135 observations and 5 variables:
\describe{
  \item{CICLO}{Days in the soybean cycle.}
  \item{GEN}{The column with the name of the genotype.}
  \item{TMED}{The column with the average temperature values.}
  \item{EST}{The column with the phenological stage.}
  \item{NN}{The column with the number of nodes.}
}
}
\source{
Simulated data for use.
}
\usage{
pheno
}
\description{
Fictitious data set for estimating soybean plastochron based on
on the number of nodes
}
\keyword{datasets}
