% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{ptn}
\alias{ptn}
\title{Data: Wheat Dataset 1}
\format{
A data.frame with 360 observations and 5 variables:
\describe{
  \item{Cult}{Wheat cultivars.}
  \item{Am}{Sample identification number.}
  \item{NQ}{Falling Number.}
  \item{W}{Gluten Strength (W).}
  \item{PTN}{Grain Protein.}
}
}
\source{
Real laboratory data.
}
\usage{
ptn
}
\description{
Data set with wheat cultivars and grain rheological characters.
}
\keyword{datasets}
