% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{ptnrg}
\alias{ptnrg}
\title{Data: Wheat Dataset 2}
\format{
A data.frame with 360 observations and 5 variables:
\describe{
  \item{CULTIVAR}{Wheat cultivars.}
  \item{REP}{Repetition number.}
  \item{PTN}{Grain protein.}
  \item{RG}{Grain yield (kg ha)}
}
}
\source{
Real field data.
}
\usage{
ptnrg
}
\description{
Wheat genotype, protein and grain yield data set
}
\keyword{datasets}
