% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{reg_GP}
\alias{reg_GP}
\title{Regression Genitor Progeny}
\usage{
reg_GP(ind, Genitor, Progenie)
}
\arguments{
\item{ind}{description}

\item{Genitor}{description}

\item{Progenie}{description}
}
\description{
Estimation of Genitor x Progeny Regression
}
\references{
Falconer, D. S., & Mackay, T. F. C. (1996). Introduction to quantitative
genetics (4th ed.). Longman.
}
\author{
Willyan Junior Adorian Bandeira

Ivan Ricardo Carvalo

Murilo Vieira Loro

Leonardo Cesar Pradebon

Jose Antonio Gonzalez da Silva
}
