% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{reg_GP}
\alias{reg_GP}
\title{Regressão Genitor Progênie}
\usage{
reg_GP(ind, Genitor, Progenie)
}
\arguments{
\item{ind}{description}

\item{Genitor}{description}

\item{Progenie}{description}
}
\description{
Estimativa da Regressão Genitor x Progênie
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
