% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indqi.R
\name{rend_ind}
\alias{rend_ind}
\title{Peeling Index and Industrial Yield}
\usage{
rend_ind(GEN, NG2M, MG, MC, RG, stat = "all", ...)
}
\arguments{
\item{GEN}{The column with the name of the genotypes.}

\item{NG2M}{The column with values for the number of grains larger than 2mm.}

\item{MG}{The column with grain mass values.}

\item{MC}{The column with karyopsis mass values.}

\item{RG}{The column with the grain yield values (kg per ha).}

\item{stat}{Logical argument. Use “all” to keep all the observations or “mean”
to extract the overall average.}
}
\description{
Calculating the Hulling Index and Industrial Yield of White Oats
}
\examples{
\donttest{
library(EstimateBreed)

data("aveia")
# Calculate the industrial yield without extracting the average
with(aveia, rend_ind(GEN,NG2M,MG,MC,RG))

# Calculate the industrial yield by extracting the average per genotype
with(aveia, rend_ind(GEN,NG2M,MG,MC,RG,stat="mean"))
}
}
\author{
Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva
}
