% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restr.R
\name{restr}
\alias{restr}
\title{Restriction of witness variability}
\usage{
restr(TEST, REP, Xi, scenario = NULL, zstat = NULL)
}
\arguments{
\item{TEST}{The column with the name of the witness}

\item{REP}{The column with the replications}

\item{Xi}{The column with the observed value for a given genotype.}

\item{scenario}{Scenario to be used for the calculation. Use “original” to
do not restrict the witnesses by the mean plus or minus the standard deviations,
or “restr” to apply the restriction.}

\item{zstat}{Logical argument. Applies Z-notation normalization if “TRUE”.}
}
\description{
Method for restricting the variability of witnesses proposed by Carvalho et al.
(2023). It uses the restriction of the mean plus or minus one standard deviation.
standard deviation, which restricts variation by removing asymmetric values.
}
\references{
Carvalho, I. R., Silva, J. A. G. da, Moura, N. B., Ferreira, L. L.,
Lautenchleger, F., & Souza, V. Q. de. (2023). Methods for estimation of
genetic parameters in soybeans: An alternative to adjust residual variability.
Acta Scientiarum. Agronomy, 45, e56156.
https://doi.org/10.4025/actasciagron.v45i1.56156
}
\author{
Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva
}
