% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risco.R
\name{risco}
\alias{risco}
\title{Risk of Disease Occurrence in Soybeans}
\usage{
risco(DAY, MONTH, TEMP, RH, disease = "rust", plot = F)
}
\arguments{
\item{DAY}{The column for the day of the month.}

\item{MONTH}{The column for the month of the year (numeric value).}

\item{TEMP}{A coluna da temperatura média do ar (em ºC).}

\item{RH}{The relative humidity column (in \%).}

\item{disease}{Define the soybean disease (Standard = “rust”).}

\item{plot}{Plot a graph of the accumulation (Default is F (FALSE)).}
}
\description{
Calculation of the Risk of Disease Occurrence in Soybeans as a Function of
Variables meteorological variables (Engers et al., 2024).
}
\examples{
\donttest{
library(EstimateBreed)

# Rust Risk Prediction
data("clima")
with(clima, risco(DY, MO, TMED, RH, disease = "rust"))
}
}
\references{
de Oliveira Engers, L.B., Radons, S.Z., Henck, A.U. et al.
Evaluation of a forecasting system to facilitate decision-making for the
chemical control of Asian soybean rust. Trop. plant pathol. 49, 539–546 (2024).
https://doi.org/10.1007/s40858-024-00649-1
}
\author{
Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva
}
