% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{SEGREGAÇÃO_PADRÃO}
\alias{SEGREGAÇÃO_PADRÃO}
\title{Segregação Padrão}
\usage{
SEGREGAÇÃO_PADRÃO(MELHORAMENTO)
}
\arguments{
\item{MELHORAMENTO}{Parâmetro base para a função imprimir a tabela}
}
\description{
Tabela didática da segregação padrão por geração
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
