% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{tamef}
\alias{tamef}
\title{Effective Population Size}
\usage{
tamef(GEN, SI, NE)
}
\arguments{
\item{GEN}{The column with the name of the genotype (progeny).}

\item{SI}{The column with the number of individuals selected}

\item{NE}{Number of individuals conducted during the selection period.}
}
\description{
Estimated effective population size adapted from Morais (1997).
}
\author{
Willyan Junior Adorian Bandeira

Ivan Ricardo Carvalo

Murilo Vieira Loro

Leonardo Cesar Pradebon

Jose Antonio Gonzalez da Silva
}
