% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiental.R
\name{TEMP_BASE}
\alias{TEMP_BASE}
\title{Plotting the optimum and cardinal temperatures for crops}
\usage{
TEMP_BASE(
  DAS,
  Var,
  crop = "soybean",
  ylab = "Atributo meteorológico",
  xlab = "Dias após Semeadura"
)
}
\arguments{
\item{DAS}{descrption}

\item{Var}{desc}

\item{ylab}{desc}

\item{xlab}{description}

\item{Cultura}{Soja, Milho, Trigo}
}
\description{
Utility function for plotting graphs of thermal preferences for crops
}
\author{
Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva
}
