% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiental.R
\name{TEMP_BASE}
\alias{TEMP_BASE}
\title{Plotando as temperaturas bases e ótimas para as culturas}
\usage{
TEMP_BASE(
  DAS,
  Var,
  Cultura = "Soja",
  ylab = "Atributo meteorológico",
  xlab = "Dias após Semeadura"
)
}
\arguments{
\item{DAS}{descrption}

\item{Var}{desc}

\item{Cultura}{Soja, Milho, Trigo}

\item{ylab}{desc}

\item{xlab}{description}
}
\description{
Utilitária para plotar gráficos dos preferendos térmicos para as culturas
}
\author{
Willyan Jr. A. Bandeira, Ivan R. Carvalho
}
