% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtgerm.R
\name{tindex}
\alias{tindex}
\title{AIC e multitrait de sementes de aveia preta}
\usage{
tindex(fc, germ, sdm, sl, radl, index = "PI")
}
\description{
Índices de seleção multicaracterística
}
\references{
Moura, N. B., Carvalho, I. R., Silva, J. A. G., Loro, M. V., Barbosa, M. H.,
Lautenchleger, F., Marchioro, V. S., & Souza, V. Q. (2021). Akaike criteria
and selection of physiological multi-character indexes for the production of
black oat seeds. Current Plant Studies, 11, 1–8.
https://doi.org/10.26814/cps2021003
}
