% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{transgressivos}
\alias{transgressivos}
\title{Seleção pelo Diferencial de Seleção (Média e Desvios)}
\usage{
transgressivos(Gen, Var, Testemunha, ylab = "Selection", xlab = "Genotypes")
}
\arguments{
\item{Gen}{The column with the genotype name}

\item{Var}{The column with the variable of interest}

\item{Witness}{The column with the value of the variable 'X' for the witnesses}
}
\description{
Selection of Transgressive Genotypes - Selection Differential (SD)
}
\author{
Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva
}
