% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecao.R
\name{transgressivos}
\alias{transgressivos}
\title{Seleção de Genótipos Transgressívos - Diferencial de Seleção (DS)}
\usage{
transgressivos(Gen, Var, Testemunha, ylab = "Seleção", xlab = "Genótipos")
}
\arguments{
\item{Gen}{desc}

\item{Var}{desc}

\item{Testemunha}{desc}
}
\description{
Seleção de Genótipos Transgressívos - Diferencial de Seleção (DS)
}
