% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{trigo}
\alias{trigo}
\title{Data: Wheat Dataset 3}
\format{
A data.frame with 19 observations and 6 variables:
\describe{
  \item{TEST}{Treatment identification.}
  \item{CE}{Ear length.}
  \item{ME}{Ear mass}
  \item{NGE}{Number of grains on the cob.}
  \item{MGE}{Grain mass of ear.}
  \item{NEE}{Number of spikelets per spike}
}
}
\source{
Real field data for use.
}
\usage{
trigo
}
\description{
Data set from a wheat experiment with different herbicide management.
}
\keyword{datasets}
