% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{vig}
\alias{vig}
\title{Data Set for Seed Vigor Extraction}
\format{
A data.frame with 54 observations and 6 variables:
\describe{
  \item{Trat}{Column with treatments.}
  \item{PC}{First Count}
  \item{G}{Germination percentage.}
  \item{CPA}{Length of aerial part.}
  \item{RAD}{Root length.}
  \item{MS}{Seedling dry mass.}
  \item{EC}{See what EC is.}
}
}
\source{
Real field data for use.
}
\usage{
vig
}
\description{
Data set from experiment with wheat genotypes subjected to different
sowing density.
}
\keyword{datasets}
