% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vig.R
\docType{data}
\name{vig}
\alias{vig}
\title{Conjunto de Dados para Extração de Vigor de Sementes}
\format{
Um data.frame com 54 observações e 6 variáveis:
\describe{
  \item{Trat}{Coluna com os tratamentos.}
  \item{PC}{Ver o que é.}
  \item{G}{Percentual de germinação.}
  \item{CPA}{Comprimento de parte aérea.}
  \item{RAD}{Comprimento radicular.}
  \item{MS}{Massa seca de plântula.}
  \item{EC}{Ver o que é EC.}
}
}
\source{
Dados reais de campo para utilização.
}
\usage{
vig
}
\description{
Conjunto de dados de experimento com genótipos de trigo submetidos a diferentes
densidade de semeadura.
}
\keyword{datasets}
